ps1: Check if in git repo more efficiently
Calling the PowerShell function Get-GitDirectory is more efficient than
calling "git rev-parse", since the latter has to spin up a new process
and the former doesn't. Also, since Get-GitDirectory is so fast (and is
basically a simplified version of what posh-git uses), there's no need
to default to Write-VcsStatus once posh-git is loaded.