Replace tabe with e in key bindings
I think if I want it in a new tab, I can just create a new tab first.
This way running multiple of these doesn't result in a bunch of tabs.

Another possibility is to create a special tab page specifically for
browsing config/plugins. If that tab page doesn't exist, this will
create it. If it does exist, it will switch to it. Or I could use a
floating window. These are future problems, though.