Reimplement vim-system-copy
That plugin was causing me some problems. The main one was that it used
command line copy/paste tools rather than the * register, and those
tools weren't always installed by default on every platform. For
example, I had to manually compile some C# I found online for paste on
Windows.

I reimplemented it in Lua and did a little bit of cleanup (e.g. the
given blockwise mode was never actually used I don't think). There's
some things in it that don't seem to work (and I'm not sure worked in
vim-system-copy either) like copying a count of lines, but that's not
super important.