Use `buffer` rather than `edit` to open scratch buffer
For some reason when I was testing it before, using `buffer` didn't add
it to the list. The documentation for 'buflisted' does say the `buffer`
command won't set it, so I assumed I needed to use `edit`. But I just
realized that using `edit` will clear options like 'buftype' and
'bufhidden', which definitely shouldn't be cleared.

Now it seems the `buffer` command sets 'buflisted' properly, so I'm just
switching back to that. I must not have sourced something properly.