Properly set 'conceallevel' in help files
The help file ftplugin was often resetting 'conceallevel' after the
autocmd. Use `vim.defer_fn` to wait until the ftplugin has run, and only
then set the option.

The more accepted way of doing this is to specify your own ftplugin file
that runs after the built-in one, but this was easier.

For some reason b:did_ftplugin wasn't being set (I'd prefer to use that
rather than 'conceallevel' itself), which I assume is because the
ftplugin loading phase had ended.