Support local configuration
The way I'm doing this is pretty simple. Basically each configuration
file is a Lua script. This script is loaded with a (currently) empty
environment to avoid the security issue of loading arbitrary Lua code on
startup.

Each config file returns config values detailing the behavior of the
global vim configuration. Currently the configuration options that are
supported are disabling editorconfig and using the neovim Lua language
server. More complex stuff can be added later as needed.