Prompt user whether they want to load local configs
Currently this doesn't really work in neovide, since it won't redraw the
prompt during startup. Even if I manually print something and call
"redraw!" it won't show up (the tildes and the blank file do, though). I
tried a bunch of stuff to get it to work, but gave up. Maybe it's a bug
that's fixed in more recent versions.

Either way, it does work in the terminal UI, so I'll just use that for
now.