Fix inactive statusline after closing api explorer
When the api explorer window was closed using a method like <c-w>c, the
WinClosed handler would be triggered to clean up any state. However, it
seems the window was actually still valid during that callback and thus
we were still calling close on the window, which resulted I guess in the
now focused window not receiving events in such a way that the
statusline became activated.

Maybe the WinEnter/BufWinEnter events weren't showing up or WinLeave was
coming after them. I'm not sure, but this fixes it :)