Set all markdown settings a second after loading
Basically the `CursorHold` autocmds weren't working because my
`defer_fn` kept running in the background and I guess blocking vim from
thinking that the cursor was actually holding. The issue is that the
markdown FileType autocmd was called multiple times, and the second time
the deferred function kept running because of its termination condition.

Basically I need something better than this for overriding file type
plugins.... vim-markdown uses an after plugin so that isn't really an
option, since it might still preempt me. I'd really like to figure out
when all filetype plugins stop running, but I'm not sure how to do that.

This is fine for now, though.