Require less escaping in Grep
This includes a few changes.

First, I changed the `Grep` command to only accept 1 argument. This way
I don't need to deal with stuff like escaping spaces in search strings.

Next, I got rid of the call to `escapecmd` in `grep`. This call would
complain on a whole bunch of characters passed in. Since its only
purpose is for grepping `<cword>`, I just made the expansion to
`<cword>` explicit before `grep` is called.

Finally, I called the `grep!` command directly rather than through
`execute`. The `execute` call was a leftover from when this was all in
VimL, but it isn't necessary anymore in Lua.