Simplify opfunc mappings
In trying to use Lua functions in the opfunc mappings, I found that we
can only get the angle bracket marks if we actually switched away from
visual mode. This is why we need to use `:lua` in the expr mappings
rather than `<cmd>lua`. Using a Lua function behaves like `<cmd>lua` in
this regard.

It turns out that the square bracket marks are set properly if you
execute g@ in visual mode, so I was over-complicating things. Basically
all my opfunc code is just modified from what I learned in
vim-system-copy. I had already improved on how that plugin did things,
so this is just the next step I guess.