Remap L to $ and g$
I don't know why I felt the need to map it to the last non-blank
character of the line before.... That isn't super useful, and also maybe
not possible for wrapped lines. Maybe it was for symmetry with H, for
which moving to the first non-blank character is actually useful?