Easily support omni-completion in addition to nvim-cmp
nvim-cmp doesn't support calling into the defined 'omnifunc', which
causes issues for plugins that use omni-completion like vimwiki. It
would really be nice to support omni-completion in nvim-cmp, or to use a
different auto-completion framework that does use omni-completion so
that we're using accepted vim concepts that will work across plugins.

For now, though, I'll just make it easier to use omni-completion in
addition to nvim-cmp.