Get rid of PowerShell Set-Location behavior on start
I can't even remember why I added this? I think it was because starting
PowerShell as admin would always place me in System32, but then why did
I make such a general solution? Anyway the behavior I prefer now is to
always open in the current directory so that for example opening a
terminal in vim will be in vim's directory.