Open all folds the first time a buffer is opened
This happens to work in exactly the way I want it to, I think? This will
only perform the action once per buffer, the first time the buffer is
entered in a window. However, you're still able to have different
`'foldlevel'` settings per buffer even in the same window, which is
weird because `'foldlevel'` should be a window-local option? I can check
and switching between buffers in the same window shows different values
of `'foldlevel'`. I think this is because vim treats folds that are
opened or closed manually specially? I'm not really sure. But it works!

Also it allows me to have buffers open in multiple windows have
different folds, which makes sense. And for some reason it doesn't
actually open all the folds for diff windows, I assume because the
autocmd kicks off when the buffers are initially created, before the
diff happens? I'm not sure. But, again, it works!

The only maybe weird behavior is if you change the folds in a buffer
then open that buffer in a new window, it will match the folds of the
previously open buffer. But I think this is OK? I guess I can change it
if I decide it's not.

I'm glad I finally found a good solution to this that seems hacky, but
actually ends up working exactly how I want because I want it to work as
if you manually opened all the folds on buffer enter!