Replace markdown plugin with tree-sitter grammar
The plugin was sketchy as heck, and I already needed to do a bunch of
hacks to get it to work. I just learned it also messes up folds in diff
buffers, so I'm resorting to the tree-sitter one since all I need is
better highlighting than the built-in one.