Use nvim-treesitter's foldexpr rather than neovim's
In neovim version 0.9.1 the editor started freezing up when reloading a
buffer (`:e` on an open buffer) if using neovim's builtin treesitter
foldexpr. It also sometimes repros in the quickfix window, although I'm
not entirely sure why. According to
https://github.com/neovim/neovim/issues/24186, this has maybe been fixed
in 0.9.2, which isn't released yet. For now I'll use the version that
works.