Use new LSP API for defining an on-attach hook
Previously you defined this hook in the config of every LSP server.
However, now it should be handled in an `autocmd`. If it isn't, it will
overwrite the `on_attach` callback for the existing server config. This
means that for example `clangd`'s default config cannot define its user
commands like `LspClangdSwitchSourceHeader`, since those are defined in
its `on_attach`.